Avoid hidden cost of accessing thread local variables
Profiling showed a lot of "__tls_get_addr" function calls which
relates to accessing thread local variables.
Since after switching we are accessing the currentKT frequently in
postSwitchFunction, this function better be inline(although it's
static, it's better to have inline keyword to show intent).
Also, instead of refrencing the currentKT by all lines of code, this
pointer is saved into a temporary variable and further refrences
uses this temporary variable to access currentKT.
This helped to increase the contex switch performance by 10%!