Support for epoll
IOHandler now supports epoll. There is only one dedicated kThread
that takes care of polling. Epoll is being used in edge-triggered
mode and an unorderd map (IOTable) is used to keep track of
notifications and requests.

For now there are custom functions for accept, recv and send. This
needs to be updated further to support more functions and possibly
later it should be added to the shim layer.