Added immediate option to context switch
Since from the time the state of a connection is changed to
POLL_WAIT to the time it is changed to POLL_READY, the block on the
mutex can cause delays and thus missing notifications, as
a temporary solution context switch now happens without checking the
cluster ready queue and if there is no work in the local ready queue
we simply switch to the main ut and run the postswitch function, to
avoid this problem. There is still a slight chance that this
situation is created if the connection is closed and another ut
start using the same connection object immediately (with the same fd
or without).