Cleand up kThread and updated the comments
Removed the copy and assign constructors. Made some variables and
functions private and provided getters and setters for them.
threadID is now being loaded when kThread is being created.
currentUT and currentKT are now private and related functions should
be called to get access to them. Doxygen comments was added.