Added a pollCache to avoid segfaults
The pointer to PollData returning from epoll could be null if the
connection was closed (thus PollData being destructed)
before unblock was called and that could cause
segfault. PollData is now allocated as part of a queue called
PollCache and the allocated memory can be recycled later, and this
ensures that there will be no null pointers returned from epoll.