Decoupled PollData form Connection
Conneciton object should only initialize a PollData any other
actions should be done through IOHandler. Also we make sure that the
file descriptor related to a connection is only added to epoll when
the first call to wait(or block) is placed. This means accept do not
need to handle epoll_ctl sys calls which slows it down.