uThread yeilded should only continue if q is empty
Since tryPopMany was changed to make sure nonblocking pop from the
queue (means not blocking on the mutex), if kThread does not receive
any uThreads form this call, it can mean both size==0 or lock was
not available. In this case kThread should only allow the current
uThread to continue if size==0, otherwise it should switch to mainUT
and do a blocking call to get more uThreads.