#23 Removing struct alloc/dealloc from postSwitch
PostSwitchStruct is now removed and both IOHandler and BlockingQueue
rely on static functions for post susupend funciton call. Each kThread
now has a pointer to a post suspend function that is set by
uThread::suspend function. IOHandler does not allocate/deallocate
anything but the BlockingQueue still requires to allocate a pair in
order to pass the proper arguments for unlocking.