Independent Scheduler
Scheduler is now independent from kThread and Cluster. KTLocal and
KTVar objects in the scheduler file can be used to define kThread
variables. All scheduling tasks are being forwarded to the
scheduler. Scheduler.h will decide which scheduler to used based on
compile time variables which will be added later.